requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
pydantic>=2.0.0
websockets>=11.0
python-multipart>=0.0.6
aiofiles>=23.0.0
websocket-client>=1.6.0
aiohttp>=3.8.0
requests>=2.31.0
mcp>=1.0.0
psutil
filelock
structlog
deap
numba
tenacity
pyyaml
